Add pong.py
This is a simple Pygame script named "PygamePong" that implements the classic game of Pong. It allows two players to control paddles using the arrow keys and "W" and "S" keys, bouncing a ball back and forth. The game continues until one player scores a point, and then the ball is reset to the center.